Discord IDs are ulongs

+++ PressGang.Bot/DiscordOptions.cs
@@ -10,6 +10,6 @@ namespace PressGang.Bot
         /// <summary>
         /// The Discord User ID of the person who owns the bot.  This determines who gets error messages.
         /// </summary>
-        public int OwnerUserId { get; set; } 
+        public ulong OwnerUserId { get; set; } 
     }
 }
